Add files via upload
We utilized the LightGBM model to fit the monthly return data of the CSI 500 and the CSI All Market (000985.XSHG) stocks to uncover the nonlinear relationship between factors and stock returns. The selected factors mainly fall into two categories: momentum factors and fundamental factors. The dataset spans from January 2010 to February 2024.

Here, we define X and Y as follows: on a monthly basis, the label Y is calculated based on the return for the next month (i.e., \( R(\text{end of next month}) / R(\text{end of this month}) - 1 \)). Returns are sorted, with the top 30% being positive samples and the bottom 30% being negative samples. X represents the factors obtained at the end of this month. Therefore, the number of training samples is approximately equal to 60% of the product of the number of months and the number of CSI stocks.

Since we did not know the optimal settings for LightGBM's hyperparameters in advance, we set the lengths of the training set, validation set, and test set to be variable. We employed grid search to find the best values for other key LightGBM hyperparameters: 'max_depth', 'learning_rate', and 'n_estimators'.

For the prediction logic, we predict the return ranking of the stock pool on a monthly basis, then calculate the monthly predicted AUC, and take the average as the overall prediction performance.

Lastly, regarding the backtesting logic, each month we adjust the portfolio by selecting the top 10% of stocks based on the predictions, allocating equal amounts to each stock.